icy-backend/app/requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]
passlib[bcrypt]
google-auth
google-api-python-client
motor
pymongo[srv]
python-dotenv
pydantic[email]
pydantic-settings
aiohttp
email-validatorrn
python-multipart
python-jose[cryptography]
passlib[bcrypt]
google-auth
google-api-python-client
motor
pymongo[srv]
python-dotenv
pydantic
pydantic-settings
aiohttp